refactor: move parser error conversion to submodule
